[PATCH] Commit "ARM: CPU: arm926ejs: Consolidate cache routines to common file" breaks u-boot on Dreamplug
Hello,

Vagrant Cascadian asked for people to test the version of u-boot packaged
for Debian Buster.  I tested u-boot on my Dreamplug and found it was not
working correctly.  I raised a bug for Debian[1] but I have also tested
with the mainline version of u-boot and found the same issues.

The first issue is that the following commit caused u-boot to no longer
be able to access usb storage on the Dreamplug:

commit 93b283d49f933f95f3a6f40762936f454ac655a8
Author: Adam Ford <aford173@gmail.com>
Date:   Thu Aug 16 13:23:11 2018 -0500

    ARM: CPU: arm926ejs: Consolidate cache routines to common file

    Four different boards had different options for enabling cache
    that were virtually all the same.  This consolidates these
    common functions into arch/arm/cpu/arm926ejs/cache.c

    This also has the positive side-effect of enabling cache on
    the Davinci (da850) boards.

    Signed-off-by: Adam Ford <aford173@gmail.com>
    [trini: Add mach-at91 to the list of consolidations]
    Signed-off-by: Tom Rini <trini@konsulko.com>

I don't have much knowledge of ARM caching, but the following patch makes
it work again on my Dreamplug.

Gbp-Pq: Topic dreamplug
Gbp-Pq: Name Commit-ARM-CPU-arm926ejs-Consolidate-cache-routines-.patch