sitara: Don't attempt to build final firmware images
The final firmware images require access to firmware not yet present in
linux-firmware. In certain instances it will be either desirable or
required to sign the elements that are included in the final firmware
images with device specific keys, which won't be present at build time.

Build and package the firmware elements that are required to form the
firmware images at a later date.

Signed-off-by: Martyn Welch <martyn.welch@collabora.com>

Gbp-Pq: Topic sitara
Gbp-Pq: Name Don-t-attempt-to-build-final-firmware-images.patch